Resolve the conflicts of branch with main
Resolve 3 files :
+ IAdminService
+ AdminService
+ AdminController
